Move StreamWriterFactory to api, move factory code to companion object
