+++ ingestor/ingestor-default/src/main/scala/za/co/absa/hyperdrive/ingestor/implementation/writer/parquet/ParquetStreamWriter.scala
@@ -17,14 +17,19 @@
 
 package za.co.absa.hyperdrive.ingestor.implementation.writer.parquet
 
+import org.apache.commons.configuration2.Configuration
 import org.apache.commons.lang3.StringUtils
-import org.apache.hadoop.conf.Configuration
+import org.apache.logging.log4j.LogManager
 import org.apache.spark.sql.streaming.{DataStreamWriter, OutputMode, StreamingQuery, Trigger}
 import org.apache.spark.sql.{DataFrame, Row}
 import za.co.absa.hyperdrive.ingestor.api.manager.OffsetManager
-import za.co.absa.hyperdrive.ingestor.api.writer.StreamWriter
+import za.co.absa.hyperdrive.ingestor.api.writer.{StreamWriter, StreamWriterFactory}
+import za.co.absa.hyperdrive.shared.configurations.ConfigurationsKeys.ParquetStreamWriterKeys.{KEY_DESTINATION_DIRECTORY, KEY_EXTRA_CONFS_ROOT}
+import za.co.absa.hyperdrive.shared.utils.ConfigUtils.getOrThrow
 
-private[writer] class ParquetStreamWriter(destination: String, val extraConfOptions: Option[Map[String,String]]) extends StreamWriter(destination) {
+import scala.util.{Failure, Success, Try}
+
+private[writer] class ParquetStreamWriter(destination: String, val extraConfOptions: Option[Map[String, String]]) extends StreamWriter(destination) {
 
   if (StringUtils.isBlank(destination)) {
     throw new IllegalArgumentException(s"Invalid PARQUET destination: '$destination'")
@@ -56,7 +61,7 @@ private[writer] class ParquetStreamWriter(destination: String, val extraConfOpti
       .outputMode(OutputMode.Append())
   }
 
-  private def addOptions(outStream: DataStreamWriter[Row], extraConfOptions: Option[Map[String,String]]): DataStreamWriter[Row] = {
+  private def addOptions(outStream: DataStreamWriter[Row], extraConfOptions: Option[Map[String, String]]): DataStreamWriter[Row] = {
     extraConfOptions match {
       case Some(options) => options.foldLeft(outStream) {
         case (previousOutStream, (optionKey, optionValue)) => previousOutStream.option(optionKey, optionValue)
@@ -65,5 +70,48 @@ private[writer] class ParquetStreamWriter(destination: String, val extraConfOpti
     }
   }
 
-  private def configureOffsets(outStream: DataStreamWriter[Row], offsetManager: OffsetManager, configuration: Configuration): DataStreamWriter[Row] = offsetManager.configureOffsets(outStream, configuration)
+  private def configureOffsets(outStream: DataStreamWriter[Row], offsetManager: OffsetManager, configuration: org.apache.hadoop.conf.Configuration): DataStreamWriter[Row] = offsetManager.configureOffsets(outStream, configuration)
+}
+
+object ParquetStreamWriter extends StreamWriterFactory {
+
+  def apply(config: Configuration): StreamWriter = {
+    val destinationDirectory = getDestinationDirectory(config)
+    val extraOptions = getExtraOptions(config)
+
+    LogManager.getLogger.info(s"Going to create ParquetStreamWriter instance using: destination directory='$destinationDirectory', extra options='$extraOptions'")
+
+    new ParquetStreamWriter(destinationDirectory, extraOptions)
+  }
+
+  private def getDestinationDirectory(configuration: Configuration): String = getOrThrow(KEY_DESTINATION_DIRECTORY, configuration, errorMessage = s"Destination directory not found. Is '$KEY_DESTINATION_DIRECTORY' defined?")
+
+  private def getExtraOptions(configuration: Configuration): Option[Map[String, String]] = {
+    import scala.collection.JavaConverters._
+    val extraOptions = configuration.getKeys(KEY_EXTRA_CONFS_ROOT)
+      .asScala
+      .map(key => getKeyValueConf(key, configuration))
+
+    if (extraOptions.nonEmpty) {
+      Some(extraOptions.toMap)
+    } else {
+      None
+    }
+  }
+
+  private def getKeyValueConf(key: String, configuration: Configuration): (String, String) = {
+    Try(parseConf(configuration.getString(key.toString))) match {
+      case Success(keyValue) => keyValue
+      case Failure(exception) => throw new IllegalArgumentException(s"Invalid extra configuration for stream writer: '$key'", exception)
+    }
+  }
+
+  private def parseConf(option: String): (String, String) = {
+    val keyValue = option.split("=")
+    if (keyValue.length == 2) {
+      (keyValue.head.trim(), keyValue.tail.head.trim())
+    } else {
+      throw new IllegalArgumentException(s"Invalid option: '$option'")
+    }
+  }
 }
